requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
aura-sr==0.0.4
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
diffusers==0.30.3
distro==1.9.0
einops==0.8.0
filelock==3.16.1
fsspec==2024.9.0
gradio_client==1.4.0
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.12
langchain-groq==0.2.0
langsmith==0.1.136
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.10.9
packaging==24.1
pillow==11.0.0
py==1.11.0
pydantic==2.9.2
pydantic_core==2.23.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
safetensors==0.4.5
setuptools==75.2.0
sniffio==1.3.1
sympy==1.13.1
tenacity==9.0.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==12.0
zipp==3.20.2
